feat: Non-blocking API(상품, 리뷰, 추천 조회), 이벤트 기반 비동기(상품, 리뷰, 추천 추가/삭제) 리액티브 마이크로서비스로 변경

+++ microservices/review-service/src/main/java/se/magnus/microservices/core/review/services/ReviewServiceImpl.java
@@ -2,8 +2,11 @@ package se.magnus.microservices.core.review.services;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.reactivestreams.Publisher;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.web.bind.annotation.RestController;
+import reactor.core.publisher.Flux;
+import reactor.core.scheduler.Scheduler;
 import se.magnus.api.core.review.Review;
 import se.magnus.api.core.review.ReviewService;
 import se.magnus.microservices.core.review.persistence.ReviewEntity;
@@ -14,6 +17,7 @@ import se.magnus.util.http.ServiceUtil;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
+import java.util.function.Supplier;
 import java.util.stream.Collectors;
 
 @Slf4j
@@ -24,25 +28,34 @@ public class ReviewServiceImpl implements ReviewService {
     private final ServiceUtil serviceUtil;
     private final ReviewRepository reviewRepository;
     private final ReviewMapper reviewMapper;
+    private final Scheduler scheduler;
 
     @Override
-    public List<Review> getReviews(int productId) {
+    public Flux<Review> getReviews(int productId) {
         if (productId < 1) throw new InvalidInputException("Invalid productId: " + productId);
 
+        log.info("Will get reviews for product with id={}", productId);
+
+        return asyncFlux(() -> Flux.fromIterable(getByProductId(productId)));
+    }
+
+    private List<Review> getByProductId(int productId) {
         List<Review> reviews = reviewRepository.findByProductId(productId)
-                .stream()
-                .map(reviewMapper::entityToApi)
-                .collect(Collectors.toList());
+            .stream()
+            .map(reviewMapper::entityToApi)
+            .collect(Collectors.toList());
 
         reviews.forEach(r -> r.setServiceAddress(serviceUtil.getServiceAddress()));
 
         log.debug("getReviews response size: {}", reviews.size());
-
         return reviews;
     }
 
     @Override
     public Review createReview(Review body) {
+
+        if (body.getProductId() < 1) throw new InvalidInputException("Invalid productId: " + body.getProductId());
+
         try {
             ReviewEntity entity = reviewMapper.apiToEntity(body);
             ReviewEntity savedEntity = reviewRepository.save(entity);
@@ -58,7 +71,13 @@ public class ReviewServiceImpl implements ReviewService {
 
     @Override
     public void deleteReviews(int productId) {
+        if (productId < 1) throw new InvalidInputException("Invalid productId: " + productId);
+
         log.debug("deleteReviews: tries to delete reviews for the product with productId: {}", productId);
         reviewRepository.deleteAll(reviewRepository.findByProductId(productId));
+    }
+
+    private <T> Flux<T> asyncFlux(Supplier<Publisher<T>> publisherSupplier){
+        return Flux.defer(publisherSupplier).subscribeOn(scheduler);
     }
 }
